requirements.txt
albumentations==1.0.3
click==8.1.3
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.4
Pillow==9.0.0
PyYAML==6.0
scikit_image==0.18.3
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
